Create Analysis_of_data_and_usage.txt
Breakdown of data and how the data will be used to interact with the game, in addition to how the different meters will be dispalyed and what they mean for the user, i.e. high light level meter means alot of lights are on, and C02 levels corresponding to how many people are in the building etc.